added controllers and services for creating and fetching accounts

+++ repository/repo.go
@@ -11,7 +11,7 @@ type RepositoryProvider interface {
 	CreateAccount(ctx context.Context, account *models.Account) (models.Account, error)
 	GetAccount(ctx context.Context, id int64) (*models.Account, error)
 	GetAccountForUpdate(ctx context.Context, id int64) (*models.Account, error)
-	ListAccounts(ctx context.Context, limit, offset int64) ([]models.Account, error)
+	ListAccounts(ctx context.Context, limit, offset int32) ([]models.Account, error)
 	UpdateAccount(ctx context.Context, id int64, balance float64) (*models.Account, error)
 	AddAccountBalance(ctx context.Context, id int64, balance float64) (*models.Account, error)
 	DeleteAccount(ctx context.Context, id int64) error
